Convert CombatRelease
