+++ src/main/java/net/countercraft/movecraft/combat/MovecraftCombat.java
@@ -7,6 +7,8 @@ import net.countercraft.movecraft.combat.features.AADirectors;
 import net.countercraft.movecraft.combat.features.AddFiresToHitbox;
 import net.countercraft.movecraft.combat.features.AntiRadar;
 import net.countercraft.movecraft.combat.features.CannonDirectors;
+import net.countercraft.movecraft.combat.features.CombatRelease;
+import net.countercraft.movecraft.combat.features.ContactExplosives;
 import net.countercraft.movecraft.combat.features.Directors;
 import net.countercraft.movecraft.combat.features.DurabilityOverride;
 import net.countercraft.movecraft.combat.features.FireballLifespan;
@@ -18,7 +20,6 @@ import net.countercraft.movecraft.combat.features.damagetracking.DamageTracking;
 import net.countercraft.movecraft.combat.listener.*;
 import net.countercraft.movecraft.combat.localisation.I18nSupport;
 import net.countercraft.movecraft.combat.player.PlayerManager;
-import net.countercraft.movecraft.combat.status.StatusManager;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.io.File;
@@ -66,17 +67,10 @@ public final class MovecraftCombat extends JavaPlugin {
         ReImplementTNTTranslocation.load(getConfig());
         TNTTracers.load(getConfig());
 
-        Config.EnableCombatReleaseTracking = getConfig().getBoolean("EnableCombatReleaseTracking", false);
-        Config.EnableCombatReleaseKick = getConfig().getBoolean("EnableCombatReleaseKick", true);
-        Config.CombatReleaseBanLength = getConfig().getLong("CombatReleaseBanLength", 60);
-        Config.CombatReleaseScuttle = getConfig().getBoolean("CombatReleaseScuttle", true);
-
         getCommand("tracersetting").setExecutor(new TracerSettingCommand());
         getCommand("tracermode").setExecutor(new TracerModeCommand());
 
         getServer().getPluginManager().registerEvents(new CraftCollisionExplosionListener(), this);
-        getServer().getPluginManager().registerEvents(new CraftReleaseListener(), this);
-        getServer().getPluginManager().registerEvents(new CraftScuttleListener(), this);
         getServer().getPluginManager().registerEvents(new DispenseListener(), this);
         getServer().getPluginManager().registerEvents(new ExplosionListener(), this);
         getServer().getPluginManager().registerEvents(new PlayerJoinListener(), this);
@@ -84,31 +78,43 @@ public final class MovecraftCombat extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new ProjectileHitListener(), this);
         getServer().getPluginManager().registerEvents(new ProjectileLaunchListener(), this);
 
+
         var damageTracking = new DamageTracking();
         getServer().getPluginManager().registerEvents(damageTracking, this);
 
         var aaDirectors = new AADirectors();
         getServer().getPluginManager().registerEvents(aaDirectors, this);
-        aaDirectors.runTaskTimer(this, 0, 1);
+        aaDirectors.runTaskTimer(this, 0, 1); // Every tick
+
         getServer().getPluginManager().registerEvents(new AddFiresToHitbox(), this);
         getServer().getPluginManager().registerEvents(new AntiRadar(), this);
+
+        var combatRelease = new CombatRelease();
+        getServer().getPluginManager().registerEvents(combatRelease, this);
+        combatRelease.runTaskTimer(this, 0, 200); // Every 10 seconds
+
         var cannonDirectors = new CannonDirectors();
         getServer().getPluginManager().registerEvents(cannonDirectors, this);
-        cannonDirectors.runTaskTimer(this, 0, 1);
+        cannonDirectors.runTaskTimer(this, 0, 1); // Every tick
+
+        var contactExplosives = new ContactExplosives();
+        getServer().getPluginManager().registerEvents(contactExplosives, this);
+        contactExplosives.runTaskTimer(this, 0, 1); // Every tick
+
         getServer().getPluginManager().registerEvents(new DurabilityOverride(), this);
-        new FireballLifespan().runTaskTimer(this, 0, 20);      // Every 1 second
+
+        new FireballLifespan().runTaskTimer(this, 0, 20); // Every 1 second
+
         getServer().getPluginManager().registerEvents(new FireballLifespan(), this);
         getServer().getPluginManager().registerEvents(new FireballPenetration(), this);
         getServer().getPluginManager().registerEvents(new MovementTracers(), this);
         getServer().getPluginManager().registerEvents(new ReImplementTNTTranslocation(), this);
+
         var tntTracers = new TNTTracers();
         getServer().getPluginManager().registerEvents(tntTracers, this);
-        tntTracers.runTaskTimer(this, 0, 1);
+        tntTracers.runTaskTimer(this, 0, 1); // Every tick
 
         playerManager = new PlayerManager();
-
-        StatusManager statusTracking = new StatusManager();
-        statusTracking.runTaskTimer(this, 0, 200);      // Every 10 seconds
     }
 
     @Override
