+++ src/main/java/net/countercraft/movecraft/combat/config/Config.java
@@ -3,8 +3,4 @@ package net.countercraft.movecraft.combat.config;
 @Deprecated(forRemoval = true)
 public class Config {
     // Combat Releasing
-    public static boolean EnableCombatReleaseTracking = false;
-    public static boolean EnableCombatReleaseKick = true;
-    public static long CombatReleaseBanLength = 60;
-    public static boolean CombatReleaseScuttle = true;
 }
