--- src/main/java/net/countercraft/movecraft/combat/listener/CraftSinkListener.java
@@ -1,19 +1,0 @@
-package net.countercraft.movecraft.combat.listener;
-
-import net.countercraft.movecraft.combat.status.StatusManager;
-import net.countercraft.movecraft.craft.PlayerCraft;
-import net.countercraft.movecraft.events.CraftSinkEvent;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-
-
-public class CraftSinkListener implements Listener {
-    @EventHandler
-    public void sinkListener(CraftSinkEvent e) {
-        if(!(e.getCraft() instanceof PlayerCraft))
-            return;
-
-        PlayerCraft playerCraft = (PlayerCraft) e.getCraft();
-        StatusManager.getInstance().craftSunk(playerCraft);
-    }
-}
